Support sphinx >= 1.8 (#156)
* Update the math block management to sphinx >= 1.8 interface

Math nodes do not have a latex attribute anymore with sphinx >= 1.8.

Math nodes are now treated as every other node, i.e. a flag is raised
when the mode is activated, then the text (content) is written, then the
flag is down.

* Use jupyter_default_lang when language attribute is default

* Remove Sphinx >= 1.8 not supported from the README

* Update travis configuration

* [math] Sphinx < 1.8 is also supported

* [litteral_block] sphinx < 1.8 is also supported

With sphinx < 1.8, the 'language' key is not alway present. A test must
be done.

* [travis] tests are made with two version of sphinx